Changed format() -> fmt() and heavy optimization
Previously we would go through the entire string, then add the position
of curly braces to an array of positions. Then we'd loop over those positions
and build our string. This means that we were practically going over the string
twice.

Now we only go over the string once, when we find some curly braces, then
we add the string.

This leads to a substantial increase in performance (around 15% in my test-cases)